types: rename export name to element3
